my-fastapi-app/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
pymongo
psycopg2-binary
motor
asyncpg
requests
aiohttp
neo4j